feat!: Add GET /functionalities-tested endpoint
The endpoint GET /data/test_functionality is deprecated in favor of the
new one.

[#OCD-3766]
